Adding bottom tabs patch
Also upgrading to latest commit dabf6a25ab01107fc1e0464ee6a3e369d1626f97
2020-05-12:

Fixed crash when window height was less or equal to bar height

When resizing, the embedded client is being resized to (height = wh - bh),
which is (<= 0) if (wh <= bh). This generates a BadValue Error leading to a
crash.  This patch fixes that by hiding the tab bar if the window height is too
small, and also sets a min_height WM hint to prevent that situation from
happening in the first place.